functions/requirements.txt
# Flask Backend
flask>=3.1.0
flask-cors>=5.0.1
gunicorn>=21.2.0
Werkzeug>=3.1.0

# Firebase
firebase-admin>=6.7.0
google-cloud-firestore>=2.20.1
firebase-functions>=0.1.0
functions-framework>=3.8.2

# WhatsApp Business API
requests>=2.31.0
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.2
pytz>=2024.1
APScheduler>=3.10.4

# Dependencies
itsdangerous>=2.2.0
click==8.1.7
Jinja2==3.1.2
MarkupSafe==2.1.3
certifi==2023.7.22
urllib3==2.0.5
charset-normalizer==3.3.0
idna==3.4
pycparser==2.21

# Optional dependencies - install manually if needed
# cffi==1.16.0
# cryptography==41.0.4
# PyJWT==2.8.0
# cachecontrol==0.13.1
# msgpack==1.0.7

# Voice processing
# deepgram-sdk==3.10.1
# aiohttp>=3.8.5 